requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn>=0.24.0,<1.0.0
python-multipart>=0.0.6,<0.1.0
python-dotenv>=1.0.0,<2.0.0

numpy>=1.24.0,<2.0.0
opencv-python-headless>=4.8.0,<5.0.0
PyWavelets>=1.5.0,<2.0.0
Pillow>=10.0.0,<12.0.0

sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
alembic>=1.13.0,<2.0.0

python-jose[cryptography]>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
slowapi>=0.1.8,<0.2.0
celery>=5.4.0,<6.0.0
redis>=5.0.0,<6.0.0

pytest>=7.0.0
httpx>=0.24.0,<1.0.0
